+++ Source/Workers/sanitizeHtml.js
@@ -1,11 +1,13 @@
 /*global define*/
 define([
         '../Core/defined',
+        '../Core/deprecationWarning',
         '../Core/RuntimeError',
         './createTaskProcessorWorker'
     ], function(
         defined,
         RuntimeError,
+        deprecationWarning,
         createTaskProcessorWorker) {
     "use strict";
 
@@ -22,6 +24,8 @@ define([
      * @see {@link http://www.w3.org/TR/workers/|Web Workers}
      */
     var sanitizeHtml = function(html) {
+        deprecationWarning('sanitize', 'sanitize has been deprecated and will be removed in Cesium 1.10.');
+
         if (!defined(html_sanitize)) {
             /*global self,importScripts*/
             self.window = {};