+++ Specs/Widgets/InfoBox/InfoBoxViewModelSpec.js
@@ -17,26 +17,35 @@ defineSuite([
         expect(viewModel.maxHeightOffset(0)).toBeDefined();
         expect(viewModel.preprocessor).toBeUndefined();
         expect(viewModel.loadingIndicatorHtml).toBeDefined();
+        expect(viewModel.sandbox).toEqual('allow-same-origin allow-popups allow-pointer-lock allow-forms');
     });
 
-    it('allows some HTML in description', function() {
-        var value = 'Testing. <script>console.error("Scripts are disallowed by default.");</script>';
+    it('sets processedDescription when preprocessor is undefined', function() {
+        var value = 'Testing.';
         var viewModel = new InfoBoxViewModel();
         viewModel.description = value;
         expect(viewModel.processedDescription).toBe(value);
     });
 
-    it('removes script tags from HTML description when used with sanitizeHtml', function() {
-        var evilString = 'Testing. <script>console.error("Scripts are disallowed by default.");</script>';
+    it('sets prcessedDescription with preprocessor result', function() {
+        var value = 'Testing.';
         var viewModel = new InfoBoxViewModel();
-        viewModel.description = evilString;
-        waitsFor(function() {
-            return viewModel.processedDescription !== viewModel.loadingIndicatorHtml;
-        });
-        runs(function() {
-            expect(viewModel.processedDescription).toContain('Testing.');
-            expect(viewModel.processedDescription).not.toContain('script');
-        });
+        viewModel.preprocessor = function(value) {
+            return value + value;
+        };
+        viewModel.description = value;
+        expect(viewModel.processedDescription).toBe(value + value);
+    });
+
+    it('sets prcessedDescription with defaultPreprocessor result', function() {
+        var value = 'Testing.';
+        InfoBoxViewModel.defaultPreprocessor = function(value) {
+            return value + value;
+        };
+        var viewModel = new InfoBoxViewModel();
+        viewModel.description = value;
+        expect(viewModel.processedDescription).toBe(value + value);
+        InfoBoxViewModel.defaultPrerocessor = undefined;
     });
 
     it('indicates missing description', function() {
@@ -88,5 +97,27 @@ defineSuite([
         viewModel.enableCamera = false;
         viewModel.isCameraTracking = true;
         expect(viewModel.cameraIconPath).toBe(disableTrackingPath);
+    });
+
+    //deprecated tests
+    it('sets prcessedDescription with sanitizer result', function() {
+        var value = 'Testing.';
+        var viewModel = new InfoBoxViewModel();
+        viewModel.sanitizer = function(value) {
+            return value + value;
+        };
+        viewModel.descriptionRawHtml = value;
+        expect(viewModel.descriptionSanitizedHtml).toBe(value + value);
+    });
+
+    it('sets prcessedDescription with defaultSanitizer result', function() {
+        var value = 'Testing.';
+        InfoBoxViewModel.defaultSanitizer = function(value) {
+            return value + value;
+        };
+        var viewModel = new InfoBoxViewModel();
+        viewModel.descriptionRawHtml = value;
+        expect(viewModel.descriptionSanitizedHtml).toBe(value + value);
+        InfoBoxViewModel.defaultSanitizer = undefined;
     });
 });